Move the executor spawning to full-node/bin
